NB22-248 11 homework

See merge request mercury-net-bootcamp-2022/vorotnikov-ivan!8

+++ Habr.Presentation/Controllers/CommentController.cs
@@ -1,6 +1,7 @@
 ﻿using Habr.BusinessLogic.Services.Interfaces;
-using Habr.DataAccess.Entities;
-using Microsoft.AspNetCore.Http;
+using Habr.Common.Exceptions;
+using Habr.Presentation.Extensions;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Habr.Presentation.Controllers
@@ -10,6 +11,7 @@ namespace Habr.Presentation.Controllers
     public class CommentController : ControllerBase
     {
         private readonly ICommentService _commentService;
+
         public CommentController(ICommentService commentService)
         {
             _commentService = commentService;
@@ -18,56 +20,46 @@ namespace Habr.Presentation.Controllers
         [HttpGet]
         public async Task<IActionResult> GetComments()
         {
-            var response = await _commentService.GetCommentsAsync();
-            return response is null || response.Count() == 0 ? NotFound("NotFound!") : Ok(response);
+            return Ok(await _commentService.GetCommentsAsync());
         }
 
         [HttpGet("{id}")]
         public async Task<IActionResult> GetCommentsById(int id)
         {
-            try
-            {
-                return Ok(await _commentService.GetCommentByIdAsync(id));
-            }
-            catch (Exception ex)
-            {
-                return NotFound(ex.Message);
-            }
+            return Ok(await _commentService.GetCommentByIdAsync(id));
         }
 
+        [Authorize]
         [HttpPost]
-        public async Task<IActionResult> CreateComment([FromQuery] int userId, [FromQuery] string text, [FromQuery] int postId, [FromQuery] int? parentId)
+        public async Task<IActionResult> CreateComment([FromQuery] string text, [FromQuery] int postId, [FromQuery] int? parentId)
         {
-            try
+            int userId = HttpContext.User.Identity.GetAuthorizedUserId();
+
+            if (parentId is null)
             {
-                if (parentId is null)
-                {
-                    await _commentService.CreateCommentAsync(userId, postId, text);
-                }
-                else
-                {
-                    await _commentService.CreateCommentAnswerAsync(userId, text, (int)parentId, postId);
-                }
-                return StatusCode(202);
+                await _commentService.CreateCommentAsync(userId, postId, text);
             }
-            catch (Exception ex)
+            else
             {
-                return BadRequest(ex.Message);
+                await _commentService.CreateCommentAnswerAsync(userId, text, (int)parentId, postId);
             }
+            return StatusCode(202);
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteComment(int id)
         {
-            try
+            var comment = await _commentService.GetFullCommentByIdAsync(id);
+            if (comment is not null && comment.UserId == HttpContext.User.Identity.GetAuthorizedUserId())
             {
                 await _commentService.DeleteCommentAsync(id);
                 return StatusCode(204);
             }
-            catch (Exception ex)
+            else
             {
-                return NotFound(ex.Message);
+                throw new BadRequestException(Common.Resources.UserExceptionMessageResource.AccessError);
             }
         }
-    } 
+    }
 }
